removed Uint16 parsing due to inconsistencies

+++ zigbee/switch-level-color-bulb/src/controller.lua
@@ -33,7 +33,7 @@ function controller.level_handler(_, device, command)
   local ep = device:get_endpoint_for_component_id(command.component)
   local move_to_level = Level.server.commands.MoveToLevel
   local lvl = math.floor(((command.args.level * 0xFF) / 0x64) + 0.5)
-  local transition_time = dtypes.Uint16(device.preferences.transitionTime)
+  local transition_time = device.preferences.transitionTime
 
   --[[
   -- MoveToLevel(
